refactor: file queries for data folder
