+++ src/data/graphql/queries-gql.ts
